WIP - removed proxmity effect into separate class - WIP on moving handles out of bb
