+++ Assets/MixedRealityToolkit.SDK/Features/UX/Scripts/BoundingBox/Visuals.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d9ebc553905704e40bac78ce7aedecb5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
