+++ Assets/MixedRealityToolkit.SDK/Features/UX/Scripts/BoundingBox/Visuals/BoundingBoxEdges.cs
@@ -1,0 +1,15 @@
+﻿using Microsoft.MixedReality.Toolkit.UI.BoundingBoxTypes;
+using UnityEngine;
+
+
+namespace Microsoft.MixedReality.Toolkit.UI
+{
+    class BoundingBoxEdges
+    {
+
+
+        
+    }
+
+
+}
