+++ Assets/MixedRealityToolkit.SDK/Features/UX/Scripts/BoundingBox/Visuals/BoundingBoxProximityEffect.cs
@@ -1,0 +1,312 @@
+﻿using System;
+using UnityEngine;
+using Microsoft.MixedReality.Toolkit.UI.BoundingBoxTypes;
+using System.Collections.Generic;
+using Microsoft.MixedReality.Toolkit.Input;
+using System.Runtime.CompilerServices;
+
+namespace Microsoft.MixedReality.Toolkit.UI 
+{
+    [Serializable]
+    public class BoundingBoxProximityEffect 
+    {
+
+        [SerializeField]
+        [Tooltip("Determines whether proximity feature (scaling and material toggling) for bounding box handles is activated")]
+        private bool proximityEffectActive = false;
+
+        /// <summary>
+        /// Determines whether proximity feature (scaling and material toggling) for bounding box handles is activated
+        /// </summary>
+        public bool ProximityEffectActive => proximityEffectActive;
+
+
+        [SerializeField]
+        [Tooltip("How far away should the hand be from a handle before it starts scaling the handle?")]
+        [Range(0.005f, 0.2f)]
+        private float handleMediumProximity = 0.1f;
+
+        [SerializeField]
+        [Tooltip("How far away should the hand be from a handle before it activates the close-proximity scaling effect?")]
+        [Range(0.001f, 0.1f)]
+        private float handleCloseProximity = 0.03f;
+
+        [SerializeField]
+        [Tooltip("A Proximity-enabled Handle scales by this amount when a hand moves out of range. Default is 0, invisible handle.")]
+        private float farScale = 0.0f;
+
+        /// <summary>
+        /// A Proximity-enabled Handle scales by this amount when a hand moves out of range. Default is 0, invisible handle.
+        /// </summary>
+        public float FarScale => farScale;
+
+        [SerializeField]
+        [Tooltip("A Proximity-enabled Handle scales by this amount when a hand moves into the Medium Proximity range. Default is 1.0, original handle size.")]
+        private float mediumScale = 1.0f;
+
+        /// <summary>
+        /// A Proximity-enabled Handle scales by this amount when a hand moves into the Medium Proximity range. Default is 1.0, original handle size.
+        /// </summary>
+        public float MediumScale => mediumScale;
+    
+        [SerializeField]
+        [Tooltip("A Proximity-enabled Handle scales by this amount when a hand moves into the Close Proximity range. Default is 1.5, larger handle size.")]
+        private float closeScale = 1.5f;
+
+        /// <summary>
+        /// A Proximity-enabled Handle scales by this amount when a hand moves into the Close Proximity range. Default is 1.5, larger handle size
+        /// </summary>
+        public float CloseScale => closeScale;
+      
+
+        [SerializeField]
+        [Tooltip("At what rate should a Proximity-scaled Handle scale when the Hand moves from Medium proximity to Far proximity?")]
+        [Range(0.0f, 1.0f)]
+        private float farGrowRate = 0.3f;
+
+        [SerializeField]
+        [Tooltip("At what rate should a Proximity-scaled Handle scale when the Hand moves to a distance that activates Medium Scale ?")]
+        [Range(0.0f, 1.0f)]
+        private float mediumGrowRate = 0.2f;
+
+        [SerializeField]
+        [Tooltip("At what rate should a Proximity-scaled Handle scale when the Hand moves to a distance that activates Close Scale ?")]
+        [Range(0.0f, 1.0f)]
+        private float closeGrowRate = 0.3f;
+
+
+
+        /// <summary>
+        /// Container for handle references and states (including scale and rotation type handles)
+        /// </summary>
+        private class Handle
+        {
+            public Transform HandleVisual;
+            public Renderer HandleVisualRenderer;
+            public HandleType Type = HandleType.None;
+            public HandleProximityState ProximityState = HandleProximityState.FullsizeNoProximity;
+        }
+
+        private List<Handle> handles;
+
+        private HashSet<IMixedRealityPointer> proximityPointers = new HashSet<IMixedRealityPointer>();
+        private List<Vector3> proximityPoints = new List<Vector3>();
+
+        public BoundingBoxProximityEffect()
+        {
+            handles = new List<Handle>();
+        }
+
+
+        public void ClearHandles()
+        {
+            if (handles != null)
+            {
+                handles.Clear();
+            }
+
+        }
+
+        public void AddHandle(HandleType type, GameObject visual)
+        {
+            handles.Add(new Handle()
+            {
+                Type = type,
+                HandleVisual = visual.transform,
+                HandleVisualRenderer = visual.GetComponent<Renderer>(),
+            });
+        }
+
+       
+
+        public void ResetHandleProximityScale(BoundingBox boundingBox)
+        {
+            if (proximityEffectActive == false)
+            {
+                return;
+            }
+
+            for (int i = 0; i < handles.Count; ++i)
+            {
+                Handle h = handles[i];
+                if (h.ProximityState != HandleProximityState.FullsizeNoProximity)
+                {
+                    h.ProximityState = HandleProximityState.FullsizeNoProximity;
+
+                    if (h.HandleVisualRenderer != null)
+                    {
+                        h.HandleVisualRenderer.material = boundingBox.HandleMaterial;
+                    }
+
+                    ScaleHandle(boundingBox, h);
+                }
+            }
+        }
+
+
+       
+        
+
+
+        public void HandleProximityScaling(BoundingBox boundingBox, IMixedRealityPointer currentPointer, Vector3 currentBoundsExtents)
+        {
+            // early out if effect is disabled
+            if (proximityEffectActive == false)
+            {
+                return;
+            }
+
+            //only use proximity effect if nothing is being dragged or grabbed
+            if (currentPointer == null)
+            {
+                proximityPointers.Clear();
+                proximityPoints.Clear();
+
+                // Find all valid pointers
+                foreach (var inputSource in CoreServices.InputSystem.DetectedInputSources)
+                {
+                    foreach (var pointer in inputSource.Pointers)
+                    {
+                        if (pointer.IsInteractionEnabled && !proximityPointers.Contains(pointer))
+                        {
+                            proximityPointers.Add(pointer);
+                        }
+                    }
+                }
+
+                // Get the max radius possible of our current bounds plus the proximity
+                float maxRadius = Mathf.Max(Mathf.Max(currentBoundsExtents.x, currentBoundsExtents.y), currentBoundsExtents.z);
+                maxRadius *= maxRadius;
+                maxRadius += handleCloseProximity + handleMediumProximity;
+
+                // Grab points within sphere of inluence from valid pointers
+                foreach (var pointer in proximityPointers)
+                {
+                    if (IsPointWithinBounds(boundingBox.transform, pointer.Position, maxRadius))
+                    {
+                        proximityPoints.Add(pointer.Position);
+                    }
+
+                    if (IsPointWithinBounds(boundingBox.transform, pointer.Result.Details.Point, maxRadius))
+                    {
+                        proximityPoints.Add(pointer.Result.Details.Point);
+                    }
+                }
+
+                // Loop through all handles and find closest one
+                int closestHandleIdx = -1;
+                float closestDistanceSqr = float.MaxValue;
+                foreach (var point in proximityPoints)
+                {
+                    for (int i = 0; i < handles.Count; ++i)
+                    {
+                        // If handle can't be visisble, skip calculations
+                        if (!boundingBox.IsHandleTypeVisible(handles[i].Type))
+                            continue;
+
+                        // Perform comparison on sqr distance since sqrt() operation is expensive in Vector3.Distance()
+                        float sqrDistance = (handles[i].HandleVisual.position - point).sqrMagnitude;
+                        if (sqrDistance < closestDistanceSqr)
+                        {
+                            closestHandleIdx = i;
+                            closestDistanceSqr = sqrDistance;
+                        }
+                    }
+                }
+
+                // Loop through all handles and update visual state based on closest point
+                for (int i = 0; i < handles.Count; ++i)
+                {
+                    Handle h = handles[i];
+
+                    HandleProximityState newState = i == closestHandleIdx ? GetProximityState(closestDistanceSqr) : HandleProximityState.FullsizeNoProximity;
+
+                    // Only apply updates if handle is in a new state or closest handle needs to lerp scaling
+                    if (h.ProximityState != newState)
+                    {
+                        // Update and save new state
+                        h.ProximityState = newState;
+
+                        if (h.HandleVisualRenderer != null)
+                        {
+                            h.HandleVisualRenderer.material = newState == HandleProximityState.CloseProximity ? boundingBox.HandleGrabbedMaterial : boundingBox.HandleMaterial;
+                        }
+                    }
+
+                    ScaleHandle(boundingBox, h, true);
+                }
+            }
+        }
+
+
+        private void ScaleHandle(BoundingBox bb, Handle handle, bool lerp = false)
+        {
+            float handleSize = handle.Type == HandleType.Scale ? bb.ScaleHandleSize : bb.RotationHandleSize;
+            float newLocalScale = GetHandleScaling(handle, handleSize, lerp);
+            handle.HandleVisual.localScale = new Vector3(newLocalScale, newLocalScale, newLocalScale);
+        }
+
+
+        private float GetHandleScaling(Handle handle, float handleSize, bool lerp = false)
+        {
+            float targetScale = 1.0f, weight = 0.0f;
+
+            switch (handle.ProximityState)
+            {
+                case HandleProximityState.FullsizeNoProximity:
+                    targetScale = farScale;
+                    weight = lerp ? farGrowRate : 1.0f;
+                    break;
+                case HandleProximityState.MediumProximity:
+                    targetScale = mediumScale;
+                    weight = lerp ? mediumGrowRate : 1.0f;
+                    break;
+                case HandleProximityState.CloseProximity:
+                    targetScale = closeScale;
+                    weight = lerp ? closeGrowRate : 1.0f;
+                    break;
+            }
+
+            return (handle.HandleVisual.localScale.x * (1.0f - weight)) + (handleSize * targetScale * weight);
+        }
+
+
+        /// <summary>
+        /// Determine if passed point is within sphere of radius around this GameObject
+        /// To avoid function overhead, request compiler to inline this function since repeatedly called every Update() for every pointer position and result
+        /// </summary>
+        /// <param name="point">world space position</param>
+        /// <param name="radiusSqr">radius of sphere in distance squared for faster comparison</param>
+        /// <returns>true if point is within sphere</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        private bool IsPointWithinBounds(Transform transform, Vector3 point, float radiusSqr)
+        {
+            // TODO
+            return (transform.position - point).sqrMagnitude < radiusSqr;
+        }
+
+
+
+
+        /// <summary>
+        /// Get the ProximityState value based on the distanced provided
+        /// </summary>
+        /// <param name="sqrDistance">distance squared in proximity in meters</param>
+        /// <returns>HandleProximityState for given distance</returns>
+        private HandleProximityState GetProximityState(float sqrDistance)
+        {
+            if (sqrDistance < handleCloseProximity * handleCloseProximity)
+            {
+                return HandleProximityState.CloseProximity;
+            }
+            else if (sqrDistance < handleMediumProximity * handleMediumProximity)
+            {
+                return HandleProximityState.MediumProximity;
+            }
+            else
+            {
+                return HandleProximityState.FullsizeNoProximity;
+            }
+        }
+    }
+}
