+++ Assets/MixedRealityToolkit.SDK/Features/UX/Scripts/BoundingBox/Visuals/BoundingBoxProximityEffect.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5356ac11b7714fa4a81cb3e093cb3d9b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
